Floating point exception solved. It was caused because I was dividing by zero in draw_line func
Need to add new file called rotate.c where theres a function to rotate every axis

The copy of the map you created must be modified along with the real mone when you translate and when you zoom. Keep that in mind for future bugs

Currently the rotation is done from the point (0, 0), which is a problem because the map escapes from the screen never to be seen again. Try to calculate the map center and then apply the difference after the rotation to accomplish the rotation from the center

+++ src/parse.c
@@ -6,7 +6,7 @@
 /*   By: jenavarr <jenavarr@student.42barcel>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/03 19:30:41 by jenavarr          #+#    #+#             */
-/*   Updated: 2023/05/24 14:55:40 by jenavarr         ###   ########.fr       */
+/*   Updated: 2023/06/13 19:19:16 by jenavarr         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -44,9 +44,9 @@ int	compose_color(int value, int endian)
 	return (color);
 }
 
-double	deg_to_rad(float degrees)
+float	deg_to_rad(float degrees)
 {
-	double	radians;
+	float	radians;
 
 	radians = degrees * (M_PI / 180.0);
 	return (radians);
